Modifer (#1)
* commit 1

* pre-vaca push

* untested pydmr modifier code

* WIP

* Moving files the modifier code does not use to 'retired.'

* Comments edited

* Moved code to 'retired.'

* Fixed the UUM-S record search feature of find_collections.py

* More retirement of code.

* Added a test for cmr.py and a comment to errLog.py

* More unit tests for the CMR module

* Unit tests for the JSON processors started - WIP

* Unit tests for the JSON processors started - WIP - 2

* More JSON processor function unit tests

* url splitting

* Added data to enable testing the different parts of the modifier

In particular, these data files should enable testing the code
without waiting on CMR over and over since they are essentially
the responses we'd get from the CMR calls in the query_cmr()
function.

* Added get_collection_granule_ids() to cmr.py

This function returns just the granule IDs in a list and can limit
the amount of data it asks for so that query times are kept low.

* url build function

* Updated cmr.py to provide better support for this code

Modified cmr_query() to use the new code in cmr.py
Updated the test for the collection_granule_list() JSON
processor used by process_request_list()

* Added documentation to query_cmr() along with type information

* test_url fct add

* SIT opendap s3 bucket

* turned off some print()

* correct template string

* first pass at earthaccess

* earthaccess auth

* Added earth_access_test.py which is a hack to read DMR++ docs using earthaccess

We will need to refine this, fixing some of the assumptions my hack makes.

* success run

* removed data files

* generated reports

* unittests first pass

* refined tests

* commit msg

* pre-merge commit

---------

Co-authored-by: James Gallagher <jgallagher@opendap.org>